Renamed audio library module to audio-router
