+++ app/src/main/java/com/twilio/video/app/AudioRouterModule.kt
@@ -1,7 +1,7 @@
 package com.twilio.video.app
 
 import android.app.Application
-import com.twilio.audio_manager.AudioDeviceSelector
+import com.twilio.audio_router.AudioDeviceSelector
 import dagger.Module
 import dagger.Provides
