+++ app/src/main/java/com/twilio/video/app/ui/room/RoomActivity.java
@@ -63,8 +63,8 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 import butterknife.OnTextChanged;
 import com.google.android.material.snackbar.Snackbar;
-import com.twilio.audio_manager.AudioDevice;
-import com.twilio.audio_manager.AudioDeviceSelector;
+import com.twilio.audio_router.AudioDevice;
+import com.twilio.audio_router.AudioDeviceSelector;
 import com.twilio.video.AspectRatio;
 import com.twilio.video.CameraCapturer;
 import com.twilio.video.LocalAudioTrack;
