+++ app/src/main/java/com/twilio/video/app/ui/room/RoomViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import com.twilio.audio_manager.AudioDevice
-import com.twilio.audio_manager.AudioDeviceSelector
+import com.twilio.audio_router.AudioDevice
+import com.twilio.audio_router.AudioDeviceSelector
 import com.twilio.video.app.ui.room.RoomViewEvent.ActivateAudioDevice
 import com.twilio.video.app.ui.room.RoomViewEvent.Connect
 import com.twilio.video.app.ui.room.RoomViewEvent.DeactivateAudioDevice
