+++ audio-router/src/main/java/com/twilio/audio_router/AudioDevice.java
@@ -1,4 +1,4 @@
-package com.twilio.audio_manager;
+package com.twilio.audio_router;
 
 import androidx.annotation.NonNull;
