+++ audio-router/src/main/java/com/twilio/audio_router/BluetoothController.java
@@ -1,4 +1,4 @@
-package com.twilio.audio_manager;
+package com.twilio.audio_router;
 
 import android.bluetooth.BluetoothAdapter;
 import android.bluetooth.BluetoothClass;
