+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-include ':twilio-audio-router'
+include ':audio-router'
